T#781: Extract 17 auth/guest/token route registrations to src/server/routes.ts
Phase 2.3 of Library #102 modularization. Unblocked by T#792 (helper exports).

server.ts: 4991 → 4428 lines (-563)
New module: src/server/routes.ts (611 lines)

Routes extracted (17 total):

**Auth (3)**:
- GET  /api/auth/status
- POST /api/auth/login (incl. login rate-limit pathway + bcrypt timing-attack
  mitigation + guest-vs-owner login branching + auth_success/auth_failure
  security event logging)
- POST /api/auth/logout

**Guest admin (8 — Gorn-only)**:
- POST   /api/guests              (create)
- GET    /api/guests              (list with online presence + msg/thread counts)
- GET    /api/guests/:id          (single + activity summary)
- PATCH  /api/guests/:id          (update display_name, expires_at, disabled_at)
- PATCH  /api/guests/:id/password (owner reset)
- DELETE /api/guests/:id          (with wsBroadcast guest_deleted)
- POST   /api/guests/:id/ban      (owner OR Beast token; guest_banned event)
- POST   /api/guests/:id/unban    (owner only; guest_unbanned event)

**Beast token (4 — owner-only)**:
- POST   /api/auth/tokens
- GET    /api/auth/tokens
- DELETE /api/auth/tokens/:id
- POST   /api/auth/tokens/rotate (owner-driven rotation)

**Beast-self (2 — bearer-token-only)**:
- GET    /api/auth/me     (Spec #51 — token info read, NEVER returns hash)
- POST   /api/auth/rotate (Spec #52 — chain-aware self-rotation)

**Auth gates preserved verbatim** (verified pre-merge):
- Session-only gates: hasSessionAuth(c) && role !== 'guest' on guest CRUD
- Beast token gates: authMethod === 'token' on /api/auth/me + /api/auth/rotate
- Rate limit: max 5 attempts per IP per 15min, SQLite-persisted (T#594)
- Timing-attack mitigation: dummy bcrypt hash on guest-not-found path
- T#718 spoof-block: banned_by derived from auth context, not request body
- Spec #51/52 chain-aware rotation: replay detection in validateToken()
- Cookie flags: httpOnly, secure, sameSite='Lax', path='/'

**Helpers**:
- setRateLimit moved INTO the new module as local closure (only used by
  /api/auth/login handler)
- getRateLimit, clearRateLimit stay in server.ts (used elsewhere)
- webPresence Map passed via DI helpers (referenced by /api/guests list)
- All other helpers (createGuest/listGuests/getGuest/getGuestByUsername/
  updateGuest/deleteGuest/banGuest/unbanGuest/isGuestActive/
  recordFailedAttempt/recordSuccessfulLogin/resetGuestPassword/
  logSecurityEvent/createToken/listTokens/revokeToken/rotateToken/
  selfRotateToken/getTokenInfo/getSetting) imported directly from
  their canonical source modules
- T#792-exported session/auth/rate-limit constants + functions
  (SESSION_COOKIE_NAME, SESSION_DURATION_MS, GUEST_SESSION_DURATION_MS,
  LOGIN_RATE_LIMIT, LOGIN_RATE_WINDOW_MS, WEB_PRESENCE_TIMEOUT_MS,
  getRateLimit, clearRateLimit, generateSessionToken, isAuthenticated)
  imported from '../server.ts'

**Registration ordering** (T#783 lesson):
registerServerRoutes registers at L4176, BEFORE the SPA fallback catch-all
at L4228. Class `module-registration-must-precede-spa-catch-all` honored.

Zero logic changes. Zero auth gate drops. Zero handler body changes.

Reviewer: bertus.
